MSVC odyssey, step 73

+++ include/libflatarray/detail/set_instance_functor.hpp
@@ -44,10 +44,10 @@ public:
         std::size_t z,
         std::size_t count) :
         source(source),
-        x(static_cast<long>(x)),
-        y(static_cast<long>(y)),
-        z(static_cast<long>(z)),
-        count(static_cast<long>(count))
+        x(x),
+        y(y),
+        z(z),
+        count(count)
     {}
 
     template<long DIM_X, long DIM_Y, long DIM_Z, long INDEX>
@@ -118,10 +118,10 @@ public:
         std::size_t z,
         std::size_t count) :
         source(source),
-        x(static_cast<long>(x)),
-        y(static_cast<long>(y)),
-        z(static_cast<long>(z)),
-        count(static_cast<long>(count))
+        x(x),
+        y(y),
+        z(z),
+        count(count)
     {}
 
     template<long DIM_X, long DIM_Y, long DIM_Z, long INDEX>
